docs(README.md): clarify instructions for including static pages in generator
Changed the wording to clarify that if you want to include a page as a non-processed page, you should include the tag `static:false` in the front matter. If you want it to be processed, you should include the `static:true` tag. Also added that the `<!-- content -->` tag in templates should be an html comment.